app/requirements/base.txt
Django==3.2.18
django-psdb-engine==1.0.6
djangorestframework==3.14.0
django-cors-headers==3.14.0
django-environ==0.9.0
django-filter==23.1
django_redis==5.2.0
celery==5.2.7
redis==4.4.0
drf-spectacular==0.26.2
gunicorn==20.1.0
huggingface-hub==0.13.4
cloudinary==1.32.0
openai==0.27.4
PyJWT==2.6.0
requests==2.28.2
whitenoise==6.4.0
transformers==4.28.1
google-api-python-client==2.86.0
oauthlib==3.2.2
firebase-admin==6.1.0
django-paypal==2.0
mysqlclient==2.1.0
flower==1.2.0
boto3==1.21.8
importlib-metadata==4.13.0
python-decouple==3.8
